lstm and streamlit/lstm_model/requirements.txt
gunicorn
flask
scikit-learn
pandas
numpy
SQLAlchemy
psycopg2-binary

--extra-index-url https://download.pytorch.org/whl/cpu
torch

